requirements.txt
Django=4.2.9
django-crispy-forms>=2.1
crispy-bootstrap4=2023.1
django-allauth=0.60.1
django-wkhtmltopdf=3.4.0
more-itertools=10.2.0
djangorestframework=3.14.0

psycopg2>=2.9